[interpreter] rename interpreter to interp
for folks without autocompletion :P
