+++ mono/mini/Makefile.am.in
@@ -43,7 +43,7 @@ RUNTIME_EXECUTABLE = $(if $(BOEHM),$(top_builddir)/mono/mini/mono-boehm,$(top_bu
 
 MINI_RUNTIME = MONO_PATH=$(CLASS) $(RUNTIME_EXECUTABLE)
 TOOLS_RUNTIME = MONO_PATH=$(mcs_topdir)/class/lib/build $(top_builddir)/runtime/mono-wrapper
-INTERPRETER_RUNTIME = $(MINI_RUNTIME) --interpreter
+INTERP_RUNTIME = $(MINI_RUNTIME) --interpreter
 RUNTIME_AOTCHECK = MONO_PATH="$(CLASS)$(PLATFORM_PATH_SEPARATOR)." $(RUNTIME_EXECUTABLE)
 
 MCS = CSC_SDK_PATH_DISABLED= $(TOOLS_RUNTIME) $(CSC) -unsafe -nowarn:0162 -nologo -noconfig -r:$(CLASS)/mscorlib.dll -r:$(CLASS)/System.dll -r:$(CLASS)/System.Core.dll
@@ -388,15 +388,15 @@ endif
 endif
 
 if ENABLE_INTERPRETER
-interpreter_sources =	\
-	interpreter/hacks.h		\
-	interpreter/interp.h	\
-	interpreter/interp-internals.h	\
-	interpreter/interp.c	\
-	interpreter/mintops.h	\
-	interpreter/mintops.def	\
-	interpreter/mintops.c	\
-	interpreter/transform.c
+interp_sources =	\
+	interp/hacks.h		\
+	interp/interp.h	\
+	interp/interp-internals.h	\
+	interp/interp.c	\
+	interp/mintops.h	\
+	interp/mintops.def	\
+	interp/mintops.c	\
+	interp/transform.c
 endif
 
 if ENABLE_LLVM
@@ -628,7 +628,7 @@ os_sources = $(darwin_sources) $(posix_sources)
 monobin_platform_ldflags=-framework CoreFoundation -framework Foundation
 endif
 
-libmini_la_SOURCES = $(common_sources) $(llvm_sources) $(llvm_runtime_sources) $(interpreter_sources) $(arch_sources) $(os_sources)
+libmini_la_SOURCES = $(common_sources) $(llvm_sources) $(llvm_runtime_sources) $(interp_sources) $(arch_sources) $(os_sources)
 libmini_la_CFLAGS = $(mono_CFLAGS)
 
 libmonoboehm_2_0_la_SOURCES =
@@ -753,7 +753,7 @@ rcheck: mono $(regtests)
 	$(MINI_RUNTIME) --regression $(regtests)
 
 richeck: mono $(iregtests)
-	$(INTERPRETER_RUNTIME) --regression $(iregtests)
+	$(INTERP_RUNTIME) --regression $(iregtests)
 
 if ARM
 check-seq-points:
